Addressed review comments #2
