+++ Tasks/Common/azure-arm-rest/azure-arm-app-service.ts
@@ -1,1141 +1,437 @@
-import msRestAzure = require("./azure-arm-common");
+import msRestAzure = require('./azure-arm-common');
 import tl = require('vsts-task-lib/task');
-import util = require("util");
-import webClient = require("./webClient");
-import azureServiceClient = require("./AzureServiceClient");
-import Model = require("./azureModels");
-import Q = require("q");
+import util = require('util');
+import webClient = require('./webClient');
+import Q = require('q');
+import { 
+    AzureEndpoint,
+    AzureAppServiceConfigurationDetails
+} from './azureModels';
 
-export class AppServiceManagementClient extends azureServiceClient.ServiceClient {
-    public appService: AppService;
+import {
+    ServiceClient,
+    ToError
+} from './AzureServiceClient';
+import { Kudu } from './azure-arm-app-service-kudu';
 
-    constructor(credentials: msRestAzure.ApplicationTokenCredentials, subscriptionId: string, options?: any) {
-        super(credentials, subscriptionId);
+var parseString = require('xml2js').parseString;
 
-        if (!!options && !!options.longRunningOperationRetryTimeout) {
-            this.longRunningOperationRetryTimeout = options.longRunningOperationRetryTimeout;
+export class AzureAppService {
+    private _endpoint: AzureEndpoint;
+    private _resourceGroup: string;
+    private _name: string;
+    private _slot: string;
+    private _appKind: string;
+    public _client: ServiceClient;
+    private _appServiceConfigurationDetails: AzureAppServiceConfigurationDetails;
+    private _appServicePublishingProfile: any;
+
+    constructor(endpoint: AzureEndpoint, resourceGroup: string, name: string, slot?: string, appKind?: string) {
+        var credentials = new msRestAzure.ApplicationTokenCredentials(endpoint.servicePrincipalClientID, endpoint.tenantID, endpoint.servicePrincipalKey, 
+            endpoint.url, endpoint.environmentAuthorityUrl, endpoint.activeDirectoryResourceID, endpoint.environment.toLowerCase() == 'azurestack');
+        this._client = new ServiceClient(credentials, endpoint.subscriptionID, 30);
+        this._endpoint = endpoint;
+        this._resourceGroup = resourceGroup;
+        this._name = name;
+        this._slot = (slot && slot.toLowerCase() == 'production') ? null : slot;
+        this._appKind = appKind;
+    }
+
+    public async start() {
+        try {
+            var webRequest = new webClient.WebRequest();
+            webRequest.method = 'POST';
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            webRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{ResourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/start`, {
+                '{ResourceGroupName}': this._resourceGroup,
+                '{name}': this._name
+            }, null, '2016-08-01');
+
+            console.log(tl.loc('StartingAppService', this._getFormattedName()));
+            var response = await this._client.beginRequest(webRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
+            }
+
+            console.log(tl.loc('StartedAppService', this._getFormattedName()));
+            return response.statusCode;
         }
-
-        this.appService = new AppService(this);
-    }
-}
-
-export class AppService {
-    private _client: AppServiceManagementClient;
-
-    constructor(client) {
-        this._client = client;
+        catch(error) {
+            throw Error(tl.loc('FailedToStartAppService', this._getFormattedName(), this._client.getFormattedError(error)));
+        }
     }
 
-    public async start(resourceGroupName: string, name: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
+    public async stop() {
+        try {
+            var webRequest = new webClient.WebRequest();
+            webRequest.method = 'POST';
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            webRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{ResourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/stop`, {
+                '{ResourceGroupName}': this._resourceGroup,
+                '{name}': this._name
+            }, null, '2016-08-01');
+
+            console.log(tl.loc('StoppingAppService', this._getFormattedName()));
+            var response = await this._client.beginRequest(webRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
+            }
+
+            console.log(tl.loc('StoppedAppService', this._getFormattedName()));
+            return response.statusCode;
+        }
+        catch(error) {
+            throw Error(tl.loc('FailedToStopAppService', this._getFormattedName(), this._client.getFormattedError(error)));
+        }
+    }
+
+    public async restart() {
+        try {
+            var webRequest = new webClient.WebRequest();
+            webRequest.method = 'POST';
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            webRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{ResourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/restart`, {
+                '{ResourceGroupName}': this._resourceGroup,
+                '{name}': this._name
+            }, null, '2016-08-01');
+
+            console.log(tl.loc('RestartingAppService', this._getFormattedName()));
+            var response = await this._client.beginRequest(webRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
+            }
+
+            console.log(tl.loc('RestartedAppService', this._getFormattedName()));
+            return response.body;
+        }
+        catch(error) {
+            throw Error(tl.loc('FailedToRestartAppService', this._getFormattedName(), this._client.getFormattedError(error)));
+        }
+    }
+
+    public async swap(slotName: string, preserveVNet?: boolean) {
+        try {
+            var webRequest = new webClient.WebRequest();
+            webRequest.method = 'POST';
+            webRequest.body = JSON.stringify({
+                targetSlot: slotName,
+                preserveVnet: preserveVNet
+            });
+
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            webRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{ResourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/slotsswap`, {
+            '{ResourceGroupName}': this._resourceGroup,
+            '{name}': this._name,
+            '{slotUrl}': slotUrl
+            }, null, '2016-08-01');
+
+            console.log(tl.loc('SwappingAppServiceSlotSlots', this._name, this.getSlot(), slotName));
+            var response = await this._client.beginRequest(webRequest);
+            if(response.statusCode == 202) {
+                response= await this._client.getLongRunningOperationResult(response);
+            }
+            
+            if(response.statusCode != 200) {
+                throw ToError(response);
+            }
+
+            console.log(tl.loc('SwappedAppServiceSlotSlots', this._name, this.getSlot(), slotName));
+            return response.statusCode;
+        }
+        catch(error) {
+            throw Error(tl.loc('FailedToSwapAppServiceSlotSlots', this._name, this.getSlot(), slotName, this._client.getFormattedError(error)));
+        }
+    }
+
+    public async get(force?: boolean): Promise<AzureAppServiceConfigurationDetails> {
+        if(force || !this._appServiceConfigurationDetails) {
+            this._appServiceConfigurationDetails = await this._get();
         }
         
-        if(!callback) {
-            throw new Error(tl.loc("CallbackCannotBeNull"));
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/start`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
+        return this._appServiceConfigurationDetails;
     }
 
-    public async startSlot(resourceGroupName: string, name: string, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
+    public async monitorAppState(state: string) {
+        state = state.toLowerCase();
+        if(["running", "stopped"].indexOf(state) == -1) {
+            throw new Error(tl.loc('InvalidMonitorAppState', state));
+        }
+
+        while(true) {
+            var appDetails = await this.get(true);
+            if(appDetails && appDetails.properties && appDetails.properties["state"]) {
+                tl.debug(`App Service state: ${appDetails.properties["state"]}`)
+                if(appDetails.properties["state"].toLowerCase() == state) {
+                    tl.debug(`App Service state '${appDetails.properties["state"]}' matched with expected state '${state}'.`);
+                    console.log(tl.loc('AppServiceState', appDetails.properties["state"]));
+                    break;
+                }
+                await webClient.sleepFor(5);
+            }
+            else {
+                tl.debug('Unable to monitor app service details as the state is unknown.');
+                break;
+            }
+        }
+    }
+
+    public async getPublishingProfileWithSecrets(force?: boolean) {
+        if(force || !this._appServicePublishingProfile) {
+            this._appServicePublishingProfile = await this._getPublishingProfileWithSecrets();
+        }
+
+        return this._appServicePublishingProfile;
+    }
+
+    public async getWebDeployPublishingProfile() {
+        var publishingProfile = await this.getPublishingProfileWithSecrets();
+        var defer = Q.defer<any>();
+        parseString(publishingProfile, (error, result) => {
+            for (var index in result.publishData.publishProfile) {
+                if (result.publishData.publishProfile[index].$.publishMethod === "MSDeploy") {
+                    defer.resolve(result.publishData.publishProfile[index].$);
+                }
+            }
+            defer.reject(tl.loc('ErrorNoSuchDeployingMethodExists'));
+        });
+
+        return defer.promise;
+    }
+
+    public async pingApplication(numberOfTimes: number) {
+        numberOfTimes = numberOfTimes ? numberOfTimes : 1;
+        try {
+            var applicationUrl: string = (await this.getWebDeployPublishingProfile()).destinationAppUrl;    
+        }
+        catch(error) {
+            tl.debug(`Unable to get publishing profile for ping application. Error: ${this._client.getFormattedError(error)}`);
         }
         
-        if(!callback) {
-            throw new Error(tl.loc("CallbackCannotBeNull"));
+        if(!applicationUrl) {
+            tl.debug('Application Url not found.');
+            return;
         }
 
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() !== 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
+        tl.debug(`Ping App Service for '${numberOfTimes}' time(s).`);
+        var webRequest = new webClient.WebRequest();
+        webRequest.method = 'GET';
+        webRequest.uri = applicationUrl;
 
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
+        while(numberOfTimes > 0) {
+            try {
+                tl.debug('pausing for 5 seconds before request');
+                await webClient.sleepFor(5);
+                var response = await webClient.sendRequest(webRequest);
+
+                tl.debug(`App Service status Code: '${response.statusCode}'. Status Message: '${response.statusMessage}'`);
+            }
+            catch(error) {
+                tl.debug(`Unable to ping App Service. Error: ${this._client.getFormattedError(error)}`);
+            }
+            finally {
+                numberOfTimes -= 1;
             }
         }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/start`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
     }
 
-    public async stop(resourceGroupName: string, name: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-        
-        if(!callback) {
-            throw new Error(tl.loc("CallbackCannotBeNull"));
-        }
-
+    public async getPublishingCredentials() {
         try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
+            var httpRequest = new webClient.WebRequest();
+            httpRequest.method = 'POST';
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/config/publishingcredentials/list`,
+            {
+                '{resourceGroupName}': this._resourceGroup,
+                '{name}': this._name,
+            }, null, '2016-08-01');
+            
+            var response = await this._client.beginRequest(httpRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
             }
+
+            return response.body;
         }
         catch(error) {
-            return callback(error);
+            throw Error(tl.loc('FailedToGetAppServicePublishingCredentials', this._getFormattedName(), this._client.getFormattedError(error)));
         }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/stop`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
     }
 
-    public async stopSlot(resourceGroupName: string, name: string, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-        
-        if(!callback) {
-            throw new Error(tl.loc("CallbackCannotBeNull"));
-        }
-
+    public async getApplicationSettings() {
         try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() !== 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
+            var httpRequest = new webClient.WebRequest();
+            httpRequest.method = 'POST';
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/config/appsettings/list`,
+            {
+                '{resourceGroupName}': this._resourceGroup,
+                '{name}': this._name,
+            }, null, '2016-08-01');
+            
+            var response = await this._client.beginRequest(httpRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
             }
 
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
-            }
+            return response.body;
         }
         catch(error) {
-            return callback(error);
+            throw Error(tl.loc('FailedToGetAppServiceApplicationSettings', this._getFormattedName(), this._client.getFormattedError(error)));
         }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/stop`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
     }
 
-    public async restart(resourceGroupName: string, name: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-        
-        if(!callback) {
-            throw new Error(tl.loc("CallbackCannotBeNull"));
-        }
-
+    public async updateApplicationSettings(applicationSettings) {
         try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
+            var httpRequest = new webClient.WebRequest();
+            httpRequest.method = 'PUT';
+            httpRequest.body = JSON.stringify(applicationSettings);
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/config/appsettings`,
+            {
+                '{resourceGroupName}': this._resourceGroup,
+                '{name}': this._name,
+            }, null, '2016-08-01');
+            
+            var response = await this._client.beginRequest(httpRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
             }
+
+            return response.body;
         }
         catch(error) {
-            return callback(error);
+            throw Error(tl.loc('FailedToUpdateAppServiceApplicationSettings', this._getFormattedName(), this._client.getFormattedError(error)));
         }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        let queryParameters = [];
-        let softRestart = (options && options.softRestart !== undefined) ? options.softRestart : undefined;
-        let synchronous = (options && options.synchronous !== undefined) ? options.synchronous : undefined;
-        if (softRestart !== null && softRestart !== undefined) {
-            queryParameters.push('softRestart=' + encodeURIComponent(softRestart.toString()));
-        }
-
-        if (synchronous !== null && synchronous !== undefined) {
-            queryParameters.push('synchronous=' + encodeURIComponent(synchronous.toString()));
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/restart`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, queryParameters, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
     }
 
-    public async restartSlot(resourceGroupName: string, name: string, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-        
-        if(!callback) {
-            throw new Error(tl.loc("CallbackCannotBeNull"));
+    public async patchApplicationSettings(properties) {
+        var applicationSettings = await this.getApplicationSettings();
+        for(var key in properties) {
+            applicationSettings.properties[key] = properties[key];
         }
 
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() !== 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
+        await this.updateApplicationSettings(applicationSettings);
 
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        let queryParameters = [];
-        let softRestart = (options && options.softRestart !== undefined) ? options.softRestart : undefined;
-        let synchronous = (options && options.synchronous !== undefined) ? options.synchronous : undefined;
-        if (softRestart !== null && softRestart !== undefined) {
-            queryParameters.push('softRestart=' + encodeURIComponent(softRestart.toString()));
-        }
-
-        if (synchronous !== null && synchronous !== undefined) {
-            queryParameters.push('synchronous=' + encodeURIComponent(synchronous.toString()));
-        }
-        
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/restart`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, queryParameters, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
-    }
-
-    public swapSlotSlot(resourceGroupName: string, name: string, slotSwapEntity: any, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if(!callback) {
-            throw new Error(tl.loc("CallbackCannotBeNull"));
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-
-            if(!slot || typeof slot.valueOf() != 'string') {
-                throw new Error(tl.loc('SourceSlotNameCannotBeNull'));
-            }
-
-            if(!slotSwapEntity) {
-                throw new Error(tl.loc('SlotSwapEntityCannotBeNull'));
-            }
-
-            if(!slotSwapEntity.targetSlot || typeof slotSwapEntity.targetSlot.valueOf() != 'string' || slotSwapEntity.preserveVnet == null || typeof slotSwapEntity.preserveVnet.valueOf() != 'boolean') {
-                throw new Error(tl.loc('InvalidSlotSwapEntity', JSON.stringify(slotSwapEntity)));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{ResourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/slotsswap`,
-        {
-            '{ResourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-        httpRequest.headers = this._client.setCustomHeaders(null);
-        httpRequest.body = JSON.stringify(slotSwapEntity);
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            var deferred = Q.defer<azureServiceClient.ApiResult>();
-            var statusCode = response.statusCode;
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else if (response.statusCode != 202) {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            else {
-                this._client.getLongRunningOperationResult(response).then((operationResponse: webClient.WebResponse) => {
-                    if (operationResponse.statusCode == 200) {
-                        deferred.resolve(new azureServiceClient.ApiResult(null, operationResponse.body));
-                    }
-                    else {
-                        deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(operationResponse)));
-                    }
-                }, (error) => deferred.reject(error));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-            (error) => callback(error));
-    }
-
-    public async swapSlotWithProduction(resourceGroupName: string, name: string, slotSwapEntity: any, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if(!callback) {
-            throw new Error(tl.loc("CallbackCannotBeNull"));
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-
-            if(!slotSwapEntity) {
-                throw new Error(tl.loc('SlotSwapEntityCannotBeNull'));
-            }
-
-            if(!slotSwapEntity.targetSlot || typeof slotSwapEntity.targetSlot.valueOf() != 'string' || slotSwapEntity.preserveVnet == null || typeof slotSwapEntity.preserveVnet.valueOf() != 'boolean') {
-                throw new Error(tl.loc('InvalidSlotSwapEntity', JSON.stringify(slotSwapEntity)));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{ResourceGroupName}/providers/Microsoft.Web/sites/{name}/slotsswap`,
-        {
-            '{ResourceGroupName}': resourceGroupName,
-            '{name}': name
-        }, null, '2016-08-01');
-        httpRequest.headers = this._client.setCustomHeaders(null);
-        httpRequest.body = JSON.stringify(slotSwapEntity);
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            var deferred = Q.defer<azureServiceClient.ApiResult>();
-            var statusCode = response.statusCode;
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else if (response.statusCode != 202) {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            else {
-                response.headers['retry-after'] = 10;
-                this._client.getLongRunningOperationResult(response).then((operationResponse: webClient.WebResponse) => {
-                    if (operationResponse.statusCode == 200) {
-                        deferred.resolve(new azureServiceClient.ApiResult(null, operationResponse.body));
-                    }
-                    else {
-                        deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(operationResponse)));
-                    }
-                }, (error) => deferred.reject(error));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-            (error) => callback(error));
     }
     
-    public async get(resourceGroupName: string, name: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
+    public async getConfiguration() {
         try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
+            var httpRequest = new webClient.WebRequest();
+            httpRequest.method = 'GET';
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/config/web`,
+            {
+                '{resourceGroupName}': this._resourceGroup,
+                '{name}': this._name,
+            }, null, '2016-08-01');
+            
+            var response = await this._client.beginRequest(httpRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
             }
+
+            return response.body;
         }
         catch(error) {
-            return callback(error);
+            throw Error(tl.loc('FailedToGetAppServiceConfiguration', this._getFormattedName(), this._client.getFormattedError(error)));
         }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'GET';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
     }
 
-    public async getSlot(resourceGroupName: string, name: string, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
+    public async updateConfiguration(applicationSettings) {
         try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
+            var httpRequest = new webClient.WebRequest();
+            httpRequest.method = 'PUT';
+            httpRequest.body = JSON.stringify(applicationSettings);
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/config/web`,
+            {
+                '{resourceGroupName}': this._resourceGroup,
+                '{name}': this._name,
+            }, null, '2016-08-01');
+            
+            var response = await this._client.beginRequest(httpRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
             }
 
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
-            }
+            return response.body;
         }
         catch(error) {
-            return callback(error);
+            throw Error(tl.loc('FailedToUpdateAppServiceConfiguration', this._getFormattedName(), this._client.getFormattedError(error)));
         }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'GET';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
     }
 
-    public async listPublishingProfileXmlWithSecrets(resourceGroupName: string, name: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
+    public async patchConfiguration(properties) {
+        var applicationSettings = await this.getConfiguration();
+        for(var key in properties) {
+            applicationSettings.properties[key] = properties[key];
         }
 
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
+        await this.updateConfiguration(applicationSettings);
 
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/publishxml`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
     }
 
-    public async listPublishingProfileXmlWithSecretsSlot(resourceGroupName: string, name: string, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/publishxml`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
+    public getSlot(): string {
+        return this._slot ? this._slot : "production";
     }
 
-    public async listPublishingCredentials(resourceGroupName: string, name: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
+    public async getKuduService() {
+        var publishingCredentials = await this.getPublishingCredentials();
+        if(publishingCredentials.properties["scmUri"]) {
+            tl.setVariable(`AZURE_APP_SERVICE_KUDU_${this._name}_${this.getSlot()}_PASSWORD`, publishingCredentials.properties["publishingPassword"], true);
+            return new Kudu(publishingCredentials.properties["scmUri"], publishingCredentials.properties["publishingUserName"], publishingCredentials.properties["publishingPassword"]);
         }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
+        throw Error(tl.loc('KuduSCMDetailsAreEmpty'));
+    }
+    private async _getPublishingProfileWithSecrets() {
         try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
+            var httpRequest = new webClient.WebRequest();
+            httpRequest.method = 'POST';
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}/publishxml`,
+            {
+                '{resourceGroupName}': this._resourceGroup,
+                '{name}': this._name,
+            }, null, '2016-08-01');
+            
+            var response = await this._client.beginRequest(httpRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
             }
+
+            var publishingProfile = response.body;
+            return publishingProfile;
         }
         catch(error) {
-            return callback(error);
+            throw Error(tl.loc('FailedToGetAppServicePublishingProfile', this._getFormattedName(), this._client.getFormattedError(error)));
         }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/config/publishingcredentials/list`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
     }
 
-    public async listPublishingCredentialsSlot(resourceGroupName: string, name: string, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
+    private async _get(): Promise<AzureAppServiceConfigurationDetails> {
         try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
+            var httpRequest = new webClient.WebRequest();
+            httpRequest.method = 'GET';
+            var slotUrl: string = !!this._slot ? `/slots/${this._slot}` : '';
+            httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/${slotUrl}`,
+            {
+                '{resourceGroupName}': this._resourceGroup,
+                '{name}': this._name,
+            }, null, '2016-08-01');
+            
+            var response = await this._client.beginRequest(httpRequest);
+            if(response.statusCode != 200) {
+                throw ToError(response);
             }
 
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
-            }
+            var appDetails = response.body;
+            return appDetails as AzureAppServiceConfigurationDetails;
         }
         catch(error) {
-            return callback(error);
+            throw Error(tl.loc('FailedToGetAppServiceDetails', this._getFormattedName(), this._client.getFormattedError(error)));
         }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/config/publishingcredentials/list`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
     }
 
-    public async getConfiguration(resourceGroupName: string, name: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'GET';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/config/web`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
-    }
-
-    public async getConfigurationSlot(resourceGroupName: string, name: string, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'GET';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/config/web`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
-    }
-
-    public async updateConfiguration(resourceGroupName: string, name: string, siteConfig: any, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'PUT';
-        httpRequest.body = JSON.stringify(siteConfig);
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/config/web`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
-    }
-
-    public async updateConfigurationSlot(resourceGroupName: string, name: string, siteConfig: any, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'PUT';
-        httpRequest.body = JSON.stringify(siteConfig);
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/config/web`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
-    }
-
-    public async getApplicationSettings(resourceGroupName: string, name: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/config/appsettings/list`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
-    }
-
-    public async getApplicationSettingsSlot(resourceGroupName: string, name: string, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'POST';
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/config/appsettings/list`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
-    }
-
-    public async updateApplicationSettings(resourceGroupName: string, name: string, appSettings: any, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'PUT';
-        httpRequest.body = JSON.stringify(appSettings);
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/config/appsettings`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
-    }
-
-    public async updateApplicationSettingsSlot(resourceGroupName: string, name: string, appSettings: any, slot: string, options: any, callback: azureServiceClient.ApiCallback) {
-        if(!callback && typeof options === 'function') {
-            callback = options;
-            options = null;
-        }
-
-        if (!callback) {
-            throw new Error('callback cannot be null.');    
-        }
-
-        try {
-            this._client.isValidResourceGroupName(resourceGroupName);
-            if(!name || typeof name.valueOf() != 'string') {
-                throw new Error(tl.loc('AppNameCannotBeNull'));
-            }
-
-            if(!slot || typeof slot.valueOf() !== 'string') {
-                throw new Error(tl.loc('SlotNameCannotBeNull'));
-            }
-        }
-        catch(error) {
-            return callback(error);
-        }
-
-        var httpRequest = new webClient.WebRequest();
-        httpRequest.method = 'PUT';
-        httpRequest.body = JSON.stringify(appSettings);
-        if(options) {
-            httpRequest.headers = this._client.setCustomHeaders(options);
-        }
-
-        httpRequest.uri = this._client.getRequestUri(`//subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Web/sites/{name}/slots/{slot}/config/appsettings`,
-        {
-            '{resourceGroupName}': resourceGroupName,
-            '{name}': name,
-            '{slot}': slot
-        }, null, '2016-08-01');
-
-        this._client.beginRequest(httpRequest).then((response: webClient.WebResponse) => {
-            let deferred = Q.defer<azureServiceClient.ApiResult>();
-            if(response.statusCode == 200) {
-                deferred.resolve(new azureServiceClient.ApiResult(null, response.body));
-            }
-            else {
-                deferred.resolve(new azureServiceClient.ApiResult(azureServiceClient.ToError(response)));
-            }
-            return deferred.promise;
-        }).then((apiResult: azureServiceClient.ApiResult) => callback(apiResult.error, apiResult.result),
-        (error) => callback(error));
+    private _getFormattedName(): string {
+        return this._slot ? `${this._name}-${this._slot}` : this._name;
     }
 }
